Separate rogue queue file, and add temporary workaround for malformed rogue dates to database class while more QA testing can occur.
